refactor: [박선호] 상품 상세 페이지 폴더 위치 수정 [DIY-SALES-APP-UI 30]
